// server/src/gql-types.ts
export type StarsQueryVariables = {
  owner: string,
  name: string,
  cursor?: string | null,
};

export type StarsQuery = {
  // Lookup a given repository by the owner and repository name.
  repository:  {
    __typename: "Repository",
    // A list of users who have starred this starrable.
    stargazers:  {
      __typename: "StargazerConnection",
      // Information to aid in pagination.
      pageInfo:  {
        __typename: "PageInfo",
        // When paginating forwards, the cursor to continue.
        endCursor: string | null,
        // When paginating forwards, are there more items?
        hasNextPage: boolean,
      },
      // A list of edges.
      edges:  Array< {
        __typename: "StargazerEdge",
        // Identifies when the item was starred.
        starredAt: string,
      } | null > | null,
    },
  } | null,
};